+++ src/distributions.rs
@@ -1,9 +1,5 @@
-pub use self::exponential_weighted::ExpWeighted;
 pub use self::raw::Raw;
 pub use self::unary::Unary;
-pub use self::weighted::Weighted;
 
-mod exponential_weighted;
 mod raw;
 mod unary;
-mod weighted;