+++ src/finite_markov_chain/fast_sample.rs
@@ -2,61 +2,51 @@
 use crate::{State, StateIterator};
 use core::fmt::Debug;
 use rand::Rng;
-use rand_distr::{Float, weighted::alias_method::Weight, Distribution};
+use rand_distr::{weighted::alias_method::Weight, Distribution};
 
 // Structs
 use crate::errors::InvalidState;
-use rand_distr::{Exp1, Exp, Uniform, weighted::alias_method::WeightedIndex};
+use rand_distr::uniform::Uniform;
+use rand_distr::weighted::alias_method::WeightedIndex;
 
 // Functions
 use core::mem;
 
-/// Finite state Markov Chain in continuous time. 
+/// Finite state Markov Chain in discrete time. 
 /// 
 /// # Costs
 /// 
 /// Construction cost: O(n), n: size of the state space.
 /// Sample cost: O(1).
 #[derive(Debug, Clone)]
-pub struct ContFiniteMarkovChain<T, W, R>
+pub struct FiniteMarkovChain<T, W, R>
 where
-    W: Float + Weight,
-    Exp1: Distribution<W>,
+    W: Weight,
     Uniform<W>: Debug + Clone,
     R: Rng,
 {
     state_index: usize,
     transition_matrix: Vec<WeightedIndex<W>>,
-    transiton_clock: Vec<W>,
     state_space: Vec<T>,
     rng: R,
 }
 
-impl<T, W, R> ContFiniteMarkovChain<T, W, R>
+impl<T, W, R> FiniteMarkovChain<T, W, R>
 where
-    W: Float + Weight + std::iter::Sum<W>,
-    Exp1: Distribution<W>,
+    W: Weight,
     Uniform<W>: Debug + Clone,
     R: Rng,
 {
     #[inline]
     pub fn new(
         state_index: usize,
-        transition_weights: Vec<Vec<W>>,
+        transition_matrix: Vec<WeightedIndex<W>>,
         state_space: Vec<T>,
         rng: R,
     ) -> Self {
-        let transition_matrix: Vec<WeightedIndex<W>> = transition_weights.clone()
-            .into_iter()
-            .map(|weights| WeightedIndex::new(weights).unwrap())
-            .collect();
-        let transiton_clock: Vec<W> = transition_weights.into_iter()
-            .map(|weights| weights.into_iter().sum::<W>())
-            .collect();
-        ContFiniteMarkovChain {
+        FiniteMarkovChain {
             state_index,
             transition_matrix,
-            transiton_clock,
             state_space,
             rng,
         }
@@ -66,18 +56,11 @@ where
     fn sample_index(&mut self) -> usize {
         self.transition_matrix[self.state_index].sample(&mut self.rng)
     }
-
-    #[inline]
-    fn sample_clock(&mut self) -> W {
-        let rate = self.transiton_clock[self.state_index];
-        Exp::new(rate).unwrap().sample(&mut self.rng)
-    }
 }
 
-impl<T, W, R> State for ContFiniteMarkovChain<T, W, R>
+impl<T, W, R> State for FiniteMarkovChain<T, W, R>
 where
-    W: Float + Weight,
-    Exp1: Distribution<W>,
+    W: Weight,
     Uniform<W>: Debug + Clone,
     T: Debug + PartialEq + Clone,
     R: Rng,
@@ -109,34 +92,31 @@ where
     }
 }
 
-impl<T, W, R> Iterator for ContFiniteMarkovChain<T, W, R>
+impl<T, W, R> Iterator for FiniteMarkovChain<T, W, R>
 where
-    W: Float + Weight,
-    Exp1: Distribution<W>,
+    W: Weight,
     Uniform<W>: Debug + Clone,
     T: Debug + PartialEq + Clone,
     R: Rng,
 {
-    type Item = (W, T);
+    type Item = T;
 
     #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.state_index = self.sample_index();
-        let period = self.sample_clock();
-        self.state().cloned().map(|x| (period, x))
+        self.state().cloned()
     }
 }
 
-impl<T, W, R> StateIterator for ContFiniteMarkovChain<T, W, R>
+impl<T, W, R> StateIterator for FiniteMarkovChain<T, W, R>
 where
-    W: Float + Weight,
-    Exp1: Distribution<W>,
+    W: Weight,
     Uniform<W>: Debug + Clone,
     T: Debug + PartialEq + Clone,
     R: Rng,
 {
     #[inline]
     fn state_as_item(&self) -> Option<<Self as std::iter::Iterator>::Item> {
-        self.state().cloned().map(|x| (W::from(0.0), x))
+        self.state().cloned()
     }
 }